Add files via upload
This is a presentation made to LASPAU-sponsored group in Miami FL, 11/20/2019.  It provides a general discussion of the critical role of IT in all missions of a university, but then provides more expansive discussion of the role of IT in enabling research.  It concludes with a discussion of the HARC project to date (Nov2019), including things we've learned so far.